Remove scheduler's dependency on the database
Instead of the scheduler directly interfacing with the database, this commit
makes the scheduler use the public and admin APIs instead. This makes the
scheduler component more module, and theoretically easy to substitute out.

+++ db/schema/001_initial_schema.sql
@@ -74,7 +74,7 @@ CREATE TABLE feed (
     id character varying NOT NULL,
     system_pk BIGINT NOT NULL,
     periodic_update_enabled boolean NOT NULL,
-    periodic_update_period integer,
+    periodic_update_period BIGINT,
     config character varying NOT NULL,
 
     UNIQUE(system_pk, id)
